Refactor and Clean
As per #8 - As far as my limited testing has gone, Code is still functional. Waiting on more testing before conforming and merging with main branch.
