+++ assistant_manager/functions/dynamic/dynamic_functions.py
