+++ assistant_manager/functions/dynamic/functions_metadata.json
