+++ assistant_manager/functions/static/default_functions_metadata.json
